# Improve repeat_interleave with scalar repeat value